requirements.txt
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
diffusers==0.31.0
fsspec==2024.10.0
httpx==0.27.2
huggingface-hub==0.26.1
ipython==8.12.3
Jinja2==3.1.4
jsonschema==4.23.0
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.0.2
openai==1.52.0
packaging==24.1
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
transformers==4.45.2
urllib3==2.2.3
